refactor(env): move sink renderers from core into their boilerplates
Core now owns only the env-registry data model (types + envVarApplies) and the
generic primitives (appendToEnv, setComposeEnvironment). The per-sink rendering
policy lives with the boilerplate that owns the sink:

- shared/env.ts          -> renderDotenv          (.env)
- docker-compose/env.ts  -> composeEnvEntries,
                            serverEnvDefaults      (compose + Dockerfile)

The $ transformers import these via ../env (relative imports are allowed in
$-files). Renderer unit tests move next to the code (shared/docker-compose gain
a vitest test script + devDeps). Generated .env/compose/Dockerfile output is
unchanged (verified by generation).
