+++ boilerplates/shared/files/$.env.ts
@@ -1,4 +1,5 @@
-import { renderDotenv, type TransformerProps } from "@batijs/core";
+import type { TransformerProps } from "@batijs/core";
+import { renderDotenv } from "../env";
 
 // Sole producer of `.env`: emits every selected feature's env vars from the
 // merged registry. Returns undefined (writes no file) when nothing applies.
