+++ boilerplates/shared/package.json
@@ -7,6 +7,7 @@
   "types": "./dist/index.d.ts",
   "scripts": {
     "check-types": "tsc --noEmit",
+    "test": "vitest run",
     "build": "bati-compile-boilerplate"
   },
   "keywords": [],
@@ -15,9 +16,11 @@
   "devDependencies": {
     "@batijs/compile": "workspace:*",
     "@batijs/core": "workspace:",
+    "@batijs/features": "workspace:",
     "@types/node": "^20.19.37",
     "vike": "0.4.259",
-    "vite": "^8.0.14"
+    "vite": "^8.0.14",
+    "vitest": "^4.1.7"
   },
   "files": [
     "dist/",
